Completed Location Module

+++ fleetuplse-backend/src/main/java/com/fleetpulse/vehicle/repository/VehicleRepository.java
@@ -1,9 +1,9 @@
-package com.fleetpulse.vehicle.repository;
+    package com.fleetpulse.vehicle.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+    import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.fleetpulse.vehicle.entity.Vehicle;
+    import com.fleetpulse.vehicle.entity.Vehicle;
 
-public interface VehicleRepository extends JpaRepository<Vehicle, Integer>{
+    public interface VehicleRepository extends JpaRepository<Vehicle, Long>{
 
-}
+    }
